This should resolve the exceptions, which occur occasionally with recurring appointments ("COMException (0xED504005): You changed one of the recurrences of this item, and this instance no longer exists. Close any open items and try again.")
This fix can be done in the Generic classes, since setting a field to the default value for its type is an generic operation.

If this fix does not solve  the Problem, Marshal.ReleaseComObject() or Marshal.FinalReleaseComObject() have to be utilized.
NOTE: Releasing a ComObject is not a generic operation and should not be done in the Generic classes.